Add MeshBase::is_replicated()
I tried to avoid this as long as possible, but is_serial() isn't
enough to distinguish between ReplicatedMesh and serialized
DistributedMesh, and we do have to distinguish those when adding new
already-partitioned nodes and elements, for which the id assignment
rules depend on mesh subclass.

We could use RTTI for this but I'd like to both avoid the need for
RTTI and avoid precluding the easy addition of other MeshBase
subclasses later.