Add Eigen-based optimization for DenseVector<T>::dot().
Also, improve and shorten the l1_norm(), l2_norm(), and linfty_norm()
Eigen-based optimizations done previously.  We are now not restricted
to using the Eigen optimizations for double: they should work with any
underlying type which Eigen itself supports.

C++ bonus meme: we get to employ the obscure forms of the 'typename'
and 'template' keywords in a single line of code.

Refs #1046.