Use DistributedMesh temporary ids even in serial
It's possible for user code to operate on a serialized
DistributedMesh, but in those cases we'll still be hitting code paths
(even in MeshRefinement::_refine_elements!) that may not perfectly
synchronized in the order in which they add elements and nodes.  The
safest way to do things is to just use the normal distributed code
path and fix up ids via communication.