Allow users to control if the matrix and RHS should be zeroed out
By default, ImplicitSystem::assemble() zeroes out the matrix and
right-hand side when the method is called.  In several cases, the matrix
can be constant and the user code can choose to not form it over and
over. For this purpose we are introducing a flag that will allow us to
do so.

Closes #1231