Add UnsteadySolver::get_second_order_dot_var API and impls
This provides an interface for the user to extract the "dot" variable
corresponding to a second order variable. The second order variable
would've been added by the user while the "dot" variable would've been
added by the FirstOrderUnsteadySolver. For SecondOrderUnsteadySolvers,
there is not dot variable, but the interface allows for consistency
with the FirstOrderUnsteadySolver while using a
SecondOrderUnsteadySolver.