Add clang-specific vectorization pragmas.
This enables the hand-coded dot() and l2_norm() algorithms to
vectorize. The pragma simply tells the compiler that it is safe to
reorder floating point optimizations in the following loop.  The
l1_norm() and linfty_norm() could not be coerced into vectorizing this
way. The L\infty norm uses a reduction variable which prevents
vectorization.  Not sure what prevents the L1-norm from vectorizing,
it may be the std::abs() call in the Complex case...