GmshIO: Allow multiple lower-dimensional elements for BCs.
Previously, we were correctly reading and storing the required
lower-dimensional elements, but we were breaking out of the loop which
assigned BCs after the first iteration, under the wrong assumption
that there would only be a single match.

Also updated the documentation where the reasoning behind using
unordered_multimap was explained.

Fixes #1252.