Fix BC construction for velocity variables
For "dot" variables in DiffSystem, their boundary conditions should
be the time derivative of the boundary condition for the parent second
order variables. Currently, we can't automatically do those derivatives
so we error out in those cases. We do handle the constant case (the
derivative is just zero), so we do that when we can.