Cleaning up the VTK parts
There was a likely VTK memory leak and/or access to deleted
memory (the leak check can be done at run-time by enabling the
VTK_DEBUG_LEAKS CMake option in the build). Other changes were
replacing non-thread safe methods with thread safe methods.
These are the getter methods that pass in the data to fill
as opposed to returning a value to the data.