Use SPARSELU for linear solver for fem_system_ex3/Euler(2)Solver
BICGSTAB (the default) was nan'ing on the Jacobians for this system
using EulerSolver and Euler2Solver (I'm not sure why). GMRES currently
doesn't work in Eigen, so all that was left was SPARSELU. That solve time
is very reasonable (less than a second), but the assembly time is
atrocious, particularly for the EulerSolver and Euler2Solver cases
(see #1268). Note that this originally manifest in --enable-complex,
but BICGSTAB is also dying using real-valued calculations as well.