Restructure how fem_system_ex3 gets run
Enable overriding arguments from the command line, allowing us to run
this example in multiple ways without having to chagne the input file.
Also have it actually run the steady case as part of run.sh; it was
an option in this example program before, but we weren't actually doing
it. This restructuring is mainly to start setting up to illustrate how
to use a FirstOrderUnsteadySolver for this problem.