Only assert_valid_dof_ids that we already set
This fixes a https://github.com/idaholab/moose/issues/1500 regression
for me on 10 processors.  All my own regression tests and use cases
for adaptivity on DistributedMesh were using a single System, so MOOSE
was the first to hit an overzealous assert that only affects corner
cases (it only hit one of their adaptivity regression tests, and only
hits it on 10 or more processors...) on multi-System problems.