Don't try to second guess ghosting functors
If we are replicated, the local element/node lists are pointless
because we want to write all elements and nodes.  If we are
distributed, those lists are dangerous, because they're based on point
neighbors rather than on the user's ghosting functor settings.  We
can't hope to get remote_elem links correct unless we're exactly
copying the real ghosting status.