Massive CheckpointIO refactoring
We now use ghosting functors to determine how much needs to be saved
in each local file in a parallel write.

We now write a "header" file in parallel so we don't have to know a
priori whether we're loading a parallel or a serial file or even how
many processors the file was written with.

We now handle N-to-M restarts.

We now do all writes/reads with a single write()/read() call, in
parallel.

We now do a much bigger swath of unit tests of all this.