AbaqusIO: Fix comments and possibly skip sideset generation.
We had previously incorrectly stated that Abaqus files don't have
sidesets in them, but they can have elsets with lower-dimensional
elements in them.

BoundaryInfo::build_side_list_from_node_list() currently assumes there
are no sidesets present in the Mesh, and overwrites/adds to existing
sidesets in surprising ways if there are.  That could certainly be
fixed, but for now we simply won't allow sideset generation for Abaqus
files in that case.