Remove nullptr checks from BoundaryInfo::sync().
This makes it possible to build BoundaryMeshes from internal sidesets,
a feature which did not work previously.  The decision of whether to
add a side to the BoundaryMesh should not be governed by whether an
Elem is on a *geometric* boundary, only by the information in the
_boundary_side_id container.

The relevant changes are in the helper functions:
* BoundaryInfo::_find_id_maps().
* BoundaryInfo::add_elements().

We are also able to hoist the multimap lookup outside of the loop over
sides, since we can get all the boundary information for a given
element's sides in a single call to equal_range(). Otherwise, the
majority of the diff is just the reindentation which accompanied the
removal of an outer if-statement.