Add ifdef'd move constructor to ReferenceCounter.
Also add a simple move constructor to the RBTheta class. This is the
simplest ReferenceCounter-derived class. I used it to confirm that the
move constructors are correctly called, and the reference counts are
correctly updated using print statements and the following simple test
code:

RBTheta rbtheta1;
RBTheta rbtheta2(std::move(rbtheta1));

Note that we are using std::move() to pass the moved-in object to the
base class move ctor. std::forward<T> could be used for this as well,
but it is more verbose since you have to specify the type of the thing
being forwarded.