Fix Tri3::contains_point().
The "barycentric coordinates" part of the algorithm is copied from
http://blackpawn.com/texts/pointinpoly. This part assumes that the
point in question already lies in the plane of the three points
defining the triangle, and returns false positives when it does not,
so we have augmented it with a normalized out-of-plane test. The new
code seems to pass the unit tests suggested by @benkirk in #993, but I
have little doubt that it could still be much better.

The right approach is probably to use so-called robust geometric
predicates with adaptive-precision floating-point arithmetic, but that
invovles quite a bit more code complexity...