build_cube/square/line with ordered element ids
We do this dictionary ordering implicitly on ReplicatedMesh, and user
code (*cough*, MOOSE, *cough*) has come to notice it and depend on it.

Our choices at this point are to enforce the same id ordering with
DistributedMesh, and give users more security when they assume that
ordering, or reorder with ReplicatedMesh, and give users a
well-deserved lesson about not relying on undocumented behavior.

This patch does the former.  My magnanimity knows no bounds.