Add unit tests for Elem::which_node_am_i().
Note: this includes a test of the error handling code in debug mode,
which is currently commented out. The error handling tests work but
print a bunch of stuff to the screen that makes it look like something
failed, so I'd like to figure that out and clean it up first if
possible. This may also be something that's simpler to do with the
Google Test suite whenever we finally switch over to using that.