Add unit tests of libmesh_assert cases.
It turns out to be pretty easy to temporarily redirect libMesh::err
while running these tests to prevent undesirable screen output.  It
may also be worth coming up with a general StreamRedirector class that
does the stream redirect in its constructor and resets it in its
destructor.  Using RAII is always the safer approach when exceptions
are being thrown.

Refs #1330.