Consistently use backslash for escaping Doxygen commands.
We were previously using both '@' and '\' to escape Doxygen
commands. Although they are equivalent, IMO it's a good idea to
standardize on a single syntax, if only to make text replacing easier
later.

Also got rid of lots of overuse of the \e flag, which was localized to
certain parts of the code and did not match the overall documentation
style.