Consistently use Doxygen "\returns" markup in class documentation.
Using these keywords improves the appearance of the documentation on
the web. IMO it looks best if you also begin the sentence following
the \returns with a capital letter and end it with a period. That way,
it looks like a complete sentence in the html.

You can find likely suspects for conversion by doing:
git grep -i returns -- include/*/*.h | grep -v "\\\returns"

So far I've made it up through the include/mesh directory, going in
alphabetical order.