make_coarsening_compatible() DistributedMesh fix
Ever since we got rid of the over-ghosting behavior, it's been
possible to have an inactive parent element whose children are remote
even on its own processor.  In such cases it's hard to correctly
determine whether that element should be flagged COARSEN_INACTIVE, and
we were not doing so correctly.

This fixes a nasty https://github.com/idaholab/moose/issues/1500 test
failure for me, and I think it might also fix #1355 if I'm lucky.