Keep is_serial* well-defined
This fixes a MOOSE --distributed-mesh "regression" for me, where an
undefined is_serial_on_proc_0 value was lulling MeshSerializer into
doing nothing and thereby causing an ExodusII_IO output to miss other
processors' values.

I say "regression" with scare quotes because this was a long-standing
bug, and it was just vagaries of nasal demons that caused it to
trigger now and not earlier.  Hopefully we just fixed some other
exodiff regression test failures too.