Avoid redundant init of old(er) Transient vectors
That init seemed to be an atavism from when these vectors weren't
registered in the System container.

Also refactor the add_vector calls, to make it easy to ensure that
SERIAL-vs-GHOSTED is still correct for those vectors even after a
clear().

This should fix one of the concerns @YaqiWang had in #1376, avoid a
little redundant work, and make the code less confusing.