More robust extra_ghost_elem handling
If such an element is coarsened away, we can still sanely deal with it
as a subactive element, for a while.

This is a workaround, though; the fix for such cases is for user code
to either switch to ghosting functors or be much more careful about
the interaction between extra_ghost_elem tracking and element
coarsening or deletion.