The testEdgeBoundaryConditions() should not require XDR.
We are primarily testing whether writing followed by reading preserves
the edge boundary ids, and this can be accomplished with an ASCII file
just as well (and is probably easier to debug). Also, updated the
CLEANFILES list and reran bootstrap (included in this commit since it
was a single-line change).

Refs #1392.