Pass sanitizer flags to the compiler which is used for linking.
Setting LDFLAGS did not appear to have any effect, so I'm adding the
flag to libmesh_LDFLAGS instead.  I don't know all the nuances of the
differences between LDFLAGS/AM_LDFLAGS/libmesh_LDFLAGS [0], but they
all seem to be used by libtool --mode=link.

Note: we have to use "-Wc," to pass flags to the compiler driver
during linking, otherwise they are simply not used.  I'm not sure if
this will break Makefiles that pass libmesh_LDFLAGS directly to a
compiler/linker without first passing it through libtool, since "-Wc,"
is most likely not a valid flag in that context.  There is also
AM_LDFLAGS, but that appears to be a Makefile variable, so I don't
know if it should be modified by configure...

Refs #1396.

[0]: https://www.gnu.org/software/automake/manual/html_node/Libtool-Flags.html