Put sideset/nodeset names in CheckpointIO header
At present we need these names to be present on all processors if we
want name maps to be replicated everywhere even in early phases of
N->M restarts where M > N.

At some point in the future we might want to distribute these maps,
but for the moment it's very very unlikely that users will have so
many boundary names that scaling will be a problem.

This breaks backwards compatibility with old cpr/cpa files, and I'm
*just* about to check some cpr/cpa files into MOOSE, so after this
passes testing I'd like to sit on it until MOOSE is ready for their
next libMesh update.