Fix CheckpointIO reads on ReplicatedMesh
Which I suppose is the ultimate N->M restart, N->1, but we weren't
handling the logic for it consistently.

This fixes the failures in https://github.com/idaholab/moose/pull/9706
for me, although naturally I'll put a workaround there too so it'll be
in shape to run before this gets merged.