Use 64bit integers in XDR "header" data
Since we're upgrading the header version and have to make a bunch of
backwards-compatibility code anyway, we might as well fix the last of
the 32-bit numbers in the format while we're at it.  This should fix
binary I/O for meshes with more than 4 billion nodes.

This ought to be the last of the problems discussed in #1418