Catch exception during side assembly in RBConstruction.
Exceptions can be thrown in side_fe_reinit, e.g. due to a zero or negative
Jacobian on an element side. This is sometimes intentional, e.g. some people
use "degenerate HEX" elements when generating meshes. In order to support this
case we just print a message and skip the side assembly on this element
if an exception occurs during side_fe_reinit.