EigenSolver: added a flag indicate whether or not to
close the matrix before solve. It is useful for nonlinear eigenvalue solvers embedded in EPS
recently. The nonlinear solvers have callbacks so that we do not need to close the matrix before solve.
If we close the matrix before solve, the memory will be shrunk and the preallocation is not right any more.