Use const char *, not std::string, in PerfLog
We still support the std::string APIs, and we still support *general*
use of the old "you give us a char * and we copy it" APIs... however,
the "we copy it" inherent to either operation turns out to be really
slow.

Allocation (and usually deallocation) of std::string was previously
occurring for every logging call, and this could ironically take
longer than any fine-grained code being logged!

The new fast_push and fast_pop methods simply let the user deal with
memory management, by requiring that the supplied pointers be pointed
at long-lived memory (in most cases a string literal); we then don't
even need to parse because we can just save pointers.

The new code is 3 or 4 times faster than the old.

Keeping the old slower methods around still leaves us with 99% of that
speed improvement, though, because we can use the new methods right
away in the library proper, and other PerfLog users don't hit it as
heavily as we do.  Perhaps only I have been foolish enough to push and
pop perf_log entries in code sections where those operations will
happen tens of millions of times in every small app.