Fix check in PetscVector<Complex>::localize_to_one
If we're in parallel, but all our DoFs are on one processor for some
reason, the PetscVector code on that processor currently thinks it's
in serial and loses sync with its peers.

This fixes --enable-complex "make check" in parallel for me.  In 1D
the adaptivity_ex3 code was actually starting with a single-element
mesh.  See #1495 for details.