Add support for detecting Intel 18.
This bug recently hit an Intel 18 user whose compiler was actually
detected to be Intel 8 (a compiler from circa 2004!) due to the
overzealous pattern matching used in compiler.m4.

A larger question is whether we should being removing support for
older compilers, especially those that don't support C++11, since we
now require that in the library.