Use AM_MAINTAINER_MODE for MetaPhysicL
Thanks to @pbauman for figuring this out and explaining it to me.

This is one change we're not going to propagate upstream.

So, it turns out that git doesn't preserve timestamps (which is
reasonable, because you don't want make to hit false negatives when
figuring out what needs to be rebuilt after you checkout an old
commit), and aclocal needs to be rerun if aclocal.m4 is older than
configure.ac (which is reasonable, because that's how we get stuff
fixed in configure via aclocal.m4), and during a checkout git
typically writes aclocal.m4 before configure.ac (which is reasonable,
because alphabet).

So when we check out this libMesh branch, make immediately intuits
that we have an older aclocal.m4 with a milliseconds-newer
configure.ac, and it tries to invoke aclocal to fix that (which is
unreasonable for us, since we want even our git repo users to be able
to run without having autotools installed).

AM_MAINTAINER_MODE, despite being a horrible hack, is by far the
*least* hackish way to fix this.