We need to use SparseMatrix<Number> still...
We don't actually ever *instantiate* SparseMatrix<Real> in the
complex-valued builds, so even if we're never going to have complex
values here, we'd better plan on filling a complex-valued matrix.

And what the hey, we might actually have complex values for some
user-supplied constraint equations, so supporting that will be nice.