Work around GCC 4.9.2 issue.
This compiler doesn't like the combination of auto-declared variables
and returning unique_ptr<Derived> as unique_ptr<Base>. Newer compilers
accept this code just fine, so presumably it is OK, this change is
just to work around the older compiler's limitation.

In playing around with this, we discovered that return by move also
fixed the problem, but according to Meyers' "Effective Modern C++"
pg. 173-174, this would actually prevent compilers from applying the
Return Value Optimization (RVO), otherwise known as copy elision,
in this case.