Fix PetscDiffSolver::reinit(), rm init()/clear() from solve()
PetscDiffSolver::reinit() will actually do the right thing now. Before,
it would segfault because clear() was called in the solve so the _snes
object had been destroyed. Which is a nice segway into the second change
in this commit.

We really don't want to be init'ing and clear'ing at each solve,
we should only do things like that if we needed to reinit so we can
reuse all the bits from the solve, e.g. we're doing timestepping or
continuation with no AMR. The impact currently will be very
small, but future plans for expensive build up of mesh hierachies
to feed to PETSc for GMG will definitely necessitate this change for
any kind of reasonable performance.