Remove update() call from FEMSystem::assembly()
Nominally we need to do this, particularly in NewtonSolver, but with
PetscDiffSolver, we're not allowed to constrain the entire solution
vector (System.solution), only current_local_solution. Thus, if
we update in assembly, we wipe out the constraints we just applied
in current_local_solution. Thus, we move update() out of assembly().
Therefore, this requires calling update() on the assembly() calls in
NewtonSolver.

While the does seem to violoate the principle of least surprise, this
also seems like the cleanest solution to me. assembly() is supposed
to be agnostic to the solver, to we don't want to put
checks/setters/whatever into assembly() to decide whether or not to
update. Additionally, it's not entirely unreasonable for assembly()
to require the current_local_solution to be prepared beforehand.
Finally, one of the main points of FEMSystem is to hide a lot of these
details from the user, so while this is a behavior change, it should
be hidden from the user entirely. Indeed, in testing this fix in
grinsfem/grins#528, no surprises were found: all tests passed, including
the new test that exposed this problem.