Check for cyclic constraints.
A simple example of a cyclic constraint is if dof 1 is constrained in terms of dof 0,
and dof 0 constrained in terms of dof 1. In this case libMesh will not resolve the
constraints correctly and we will get an incorrect solution.

This change asserts that cyclic constraints are not present in debug/devel mode, and
gives us the option to check for cyclic constraints and throw an error in opt mode.