Add PetscDMWrapper::init_and_attach_petscdm
First pass at setting up PETSc DM data. This should be the only
main public method needed by the user to interact with this. Idea
is the user gives us their System and their SNES, we'll figure out
the DM hierarchy and set it up and then attach it to the SNES. Then
the user should able to now query the solver from the command. At
the current state, we're only interacting with the finest mesh
level, but we should be able to at least do recursive fieldsplits
from the command line using this interface.