Always detect compiler's OpenMP flags.
We do this regardless of what the user specified for
--with-thread-model, since if the compiler supports OpenMP, we must
maintain control over how many OpenMP threads are spawned in other
libraries. The number of OpenMP threads that get spawned depends on:

1.) The environment variable OMP_NUM_THREADS
2.) Programmatic calls to omp_set_num_threads().

If the former is unset and the latter is never called, the default is
to use "1 thread per CPU" [0].  This can lead to dramatic slowdowns in
MPI codes due to over-subscription of the hardware. That is, if you
are running on a machine with 4 cores using "mpiexec -np 4" and
relying on the default value of OMP_NUM_THREADS, there will potentially
be 16 total threads launched in OMP parallel regions.

[0]: https://gcc.gnu.org/onlinedocs/libgomp/OMP_005fNUM_005fTHREADS.html

Refs #1579.