SNESSetFromOptions --> PetscDiffSolver::solve()
We need to call SNESSetFromOption later since we need to have everything
setup for some specific options (e.g. for subsolvers in fieldsplit
preconditioning), so let's just call it right before the solve.

For the old style fieldsplit though, PETSc needs to know that we want
a fieldsplit preconditioner before we use that part of the PETSc API,
so if we're using the old style fieldsplit, we need to call
SNESSetFromOptions before we call petsc_auto_fieldsplit.