Take a better guess at which checkpoint files to read
When asked to read a checkpoint mesh, it makes sense to read the serial
checkpoint mesh if we are actually working with replicated mesh. This
fixes the cases when you are running in parallel with a replicated mesh
but also happen to have parallel checkpoints written for the same number
of files. Without this patch, it's possible to run into having the
EquationSystems checkpoint out of syntax with the mesh Checkpoint.