Use alltoall to filter out non-communicating-procs
An alltoall operation is O(P log(P), whereas naively sending all
possible messages was O(P^2).

For me on a 320 processor half-cluster run with OpenMPI, this speeds up
this stretch of code by a factor of 4, which speeds up the total
repartitioning time by 25%.  This is not a sufficient fix for #1598 but
it's a good start.

And there are several other DistributedMesh communications which would
benefit from this same idiom...