Include the right header for PetscBLASInt.
For some reason in c3659b11 I thought this type was defined in
petscblaslapack.h, but it seems to have been in petscsys.h for some
time. This should help fix an issue raised over in xikaij/fredkin-koehler#3
where conflicting BLAS functions were being defined in both scalfmm
and PETSc.

Refs #1593.