Call choose_processor_id() where appropriate
Hopefully this will reduce the imbalance we can get on nodes when
using large numbers of processors without extremely large numbers of
elements.

This refactor enforces the consistency we need between different parts
of the code which select node processor ids: the canonical heuristic
is now a Node method, and we just use that method everywhere.