Replace manual if-tests with AS_IF.
This is apparently the "official" way to do if-tests in configure
scripts. The way we have typically done these test in the past has
apparently worked, but it may not be as portable as it could otherwise
be.

https://stackoverflow.com/questions/21015573/how-to-put-logical-or-in-autoconf-m4-as-if
http://www.idryman.org/blog/2016/03/14/autoconf-tutorial-2/

Note: you have to be careful not to quote AC_MSG_RESULT, otherwise it
will be treated as a string and ./bootstrap will fail.

https://stackoverflow.com/questions/8811381/possibly-undefined-macro-ac-msg-error