Count local dof-bearing nodes in ARefEE
This is useful when experimenting with code changes to enable
different node partitioning heuristics, where if we don't keep the
heuristic consistent through all phases of uniform mesh refinement and
coarsening we can no longer properly restore cached coarse solution
vectors.

This assertion failure is closer to the root of the problem than the
PETSc local size mismatch failure that I get otherwise.