Factor out Node::choose_processor_id heuristic
We actually use this in at least 3 places and we might find a fourth
later, so let's get it all in sync.

Keep Node::choose_processor_id() in the .C file:

If we're going to make this more complicated then it's probably not
worth inlining, especially in such an often-included header.

Default to old node partitioning heuristic: for backwards
compatibility, plus preconditioner efficiency in most use cases.