Improve/generalize configure test for XDR headers.
Fedora 28 only installs the rpc.h and xdr.h headers if you install the
"libtirpc-devel" RPM; this functionality is no longer built in to
glibc. The headers are also installed in a new location,
/usr/include/tirpc/rpc instead of the old location, /usr/include/rpc,
and you have to link against a library, -ltirpc, in order to use the
XDR functions.

For more information, see: https://fedoraproject.org/wiki/Changes/SunRPCRemoval

The original code was running a mix of AC_CHECK_FUNC and
AC_CHECK_HEADER checks, and it was a bit redundant, so I changed it to
a single AC_LINK_IFELSE test.  I also moved the XDR testing from
libmesh_core_features.m4 to libmesh_optional_packages.m4 since that
made it easier to optionally append to the libmesh_optional_INCLUDES
and libmesh_optional_LIBS variables.

We now no longer have separate defines for the rpc.h and xdr.h
headers, you either have both or you have neither, and this is
controlled by the LIBMESH_HAVE_XDR define.

Refs #1679.