Fix potential overflow issue in NemesisIO.
The original line of code had an arithmetic expression involving the
product of two ints in order to compute an implied degree of freedom
index. Unfortunately, in some cases this product was large enough to
overflow, leading to a negative value which was subsequently cast to
an (unsigned) dof_id_type, leading to a huge number. Upgrading the
first operand to dof_id_type is a simple fix for the problem.

cc: @YaqiWang