Use C++11 fixed type enumerations.
Specifying a fixed type prevents the compiler from trying to optimize
by choosing the smallest possible integral type which fits the range
of enumerated values, and forces it to use the specified type instead.
See also: http://en.cppreference.com/w/cpp/language/enum

This is useful in contexts where values outside the original
enumeration range are expected to be used, for example in the Order
enumeration as discussed in #1711.