Be more explicit about some defaulted/deleted functions.
These changes currently only affect the MeshBase hierarchy and the
ExactSolution class, but eventually every class in the library should
be more specific about which of the 5 "special functions" (copy/move
ctor, copy/move assignment operator, destructor) are deleted or
defaulted.

The reason this is important is because the compiler will not
automatically generate _any_ of the 5 special functions if any _one_
of them is user-defined (this includes the destructor). For more
information, the following blog post by Scott Meyers is a good
introduction:

http://scottmeyers.blogspot.com/2014/03/a-concern-about-rule-of-zero.html

What this rule means for e.g. the MeshBase class is that, even if one were
to explicitly try to move construct one, the copy constructor was
actually being called. That is, the following code snippet:

Mesh mesh1(init.comm());
...
Mesh mesh2(std::move(mesh1));

would silently call the MeshBase copy constructor.

This PR makes it so that the default move constructor is generated for
the MeshBase hierarchy of classes and so that the code above really
calls the move constructor. This PR also explicitly deletes special
functions which cannot be defaulted, giving a reason in the docs. This
way, if a user tries to call one of the deleted functions they get an
easy-to-understand error message, and can go look up the reason in
more detail.

In summary, the following specific updates were made:
* ExactSolution
  - default move ctor
  - copy ctor deleted (shallow copy not allowed due to containers of unique_ptrs).
  - copy and move assignment operators deleted (class contains a const reference member)
* MeshBase
  - default move ctor
  - delete copy and move assignment operators (the default versions of these would leak).
    In C++03 this function was private+unimplemented, i.e. the exact scenario deleted
    functions were created to handle.