Unit test ParsedFunction copy/move ctors and move assignment operator.
There is apparently a compiler-generated copy assignment operator for
FunctionParserADBase, but this can't be used safely since that class
manually manages memory. Therefore we will delete the copy assignment
operator at the ParsedFunction level.

Note: move assignment of ParsedFunction should work even though
FunctionParserADBase objects are not themselves move-assign'able.
That is it should be OK to move-assign a vector of such objects, as
this does not require the entries in the vector to be move-assignable
themselves.