EpetraMatrix is not default copy assignable.
We were previously calling the copy assignment operator in the
get_transpose() function, but this would shallow copy the _mat, _map,
and _graph objects and theoretically lead to issues in the
destructor... but I just noticed that the destructor doesn't clean up
anything anyway. It seems unlikely that any test using EpetraMatrix
would be valgrind clean.