Add geometric constants num_nodes, num_sides, etc. to Elem classes.
Note that we have been using inline initialization of static constants
in the library for quite a while (see dof_object.h) but this may be
the first time we've used those constants to size static arrays? It
seems to work fine for the compilers I've tested, but, as described in
an SO article [0], there is also a clever alternative based on enums
that we could instead employ. Note that "declaring" the constants
twice is required, as far as I can tell, in order to prevent linker
errors.  At least you only have to specify the actual value once...

I have also introduced a std::find-based implementation for the
is_node_on_side() and is_node_on_edge() routines of the various Elem
classes which IMO is a bit cleaner than what was there
previously. This approach makes use of the C++11 std::begin() and
std::end() operators for C-style arrays, so a configure test was added
to ensure that compilers support that feature.

Finally, the new nodes_on_side() routines were also simplified using
std::begin() and std::end(), and rather than invoking the machinery of the
std::erase/remove idiom, we can just decide how much to "trim" off the
ends of the arrays for Elems with heterogeneous collections of sides.

[0]: https://stackoverflow.com/questions/8938574/using-static-const-const-as-array-bound