Error if using PetscSection and not node-madjor ordering
Currently, PetscSection assumes a node-major ordering without any detailed
offset information from the user. @knepley has a draft of a function
PetscSectionSetUseFieldOffsets that will convert the internal ordering
in the PetscSection to var-major. Preliminary testing uncovered modifications
needed on the PETSc end to support var-major ordering automatically.
@knepley is currently doing the update and will notify once it's
complete. Until then, to be on the safe side, we'll throw an error
if the user isn't using node-major ordering.