Add DofMap::local_index
I need this little method, but I think it's also useful to wrap
up the rule for determining whether a dof index is local or not.
Went ahead and used this in DofMap::semilocal_index; I didn't
search at all for any more uses, but I'd found
DofMap::semilocal_index while looking for a version of
DofMap::local_index so went ahead and removed that duplicate code.