Use member subdomain_map, don't create a new one
I didn't notice this until -Wshadow screamed at me, but honestly I
can't see why NemesisIO_Helper::subdomain_map needs to be recreated as
a local variable here.  Even the ordering looks like it's set up the
same way.

So let's try stripping the local variable out.

It looks like the only difference between the local and the member may
be the fact that the local had entries for blocks with no local
elements, but we can change the code slightly to work around that.