Prefer vector::data() to &v[0].
Calling data() on an empty vector _may_ or _may not_ return a nullptr,
hence it is still undefined behavior to attempt to use the value of
data() returned by an empty container.

One benefit of using data() is that the call itself can never fail as
the function is marked noexcept. The same is not true for operator[].
Note: this does not mean you can actually _use_ the pointer returned by data(),
you must always verify that it came from a non-empty container.

Finally, for future people reading the code it will be easier to find
documentation for a function call like data() than a "trick" like
&v[0].

https://en.cppreference.com/w/cpp/container/vector/data
https://stackoverflow.com/questions/10740094/stdvector-vec-data-or-vec0

A good regex for finding instances of this code is the following:

git grep -e "&\S*\[0" -- include/*/*.h src/*/*.C