Move possibly_receive() to MPI-independent section
The libmesh_call_mpi wrapper() should hide the actual MPI call,
and then initializing the int to 0 should make the --disable-mpi
behaviour "possibly_receive always returns false", which sounds
accurate enough.

Now I just need to test to see if this breaks horribly in the context
of the NBX parallel_sync.h when MPI is disabled.