GenericProjector fix.
I'm not sure why this never came up before, but this fixes an
assert/crash for me when trying to call System::project_solution() on
a System with FIRST, LAGRANGE variables. Apparently we don't have any
tests for that case, but JxW is definitely needed so it should be
pre-requested.