Update to RBEIMAssembly::evaluate_basis_function()
Change this function so that it takes in a set of points rather than a quadrature
rule. The new approach is more general and less error prone, e.g. the previous
approach led to problems when using separate meshes for the EIM and the RB assembly
since the quadrature rules may not be consistent, but by passing in the appropriate
points using the new approach we can now make sure that it works correctly.